requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
ascii_colors==0.11.20
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
configparser==7.2.0
cryptography==46.0.5
distro==1.9.0
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.49.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
json_repair==0.58.5
lightrag-hku==1.4.10
multidict==6.7.1
nano-vectordb==0.0.4.3
networkx==3.6.1
numpy==2.4.3
openai==2.26.0
packaging==26.0
pandas==2.3.3
pipmaster==1.1.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.2.28
requests==2.32.5
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
xlsxwriter==3.2.9
yarl==1.23.0
